paas-python-web/requirements.txt
# Place your python requirements in this document
# Refer to https://pip.pypa.io/en/stable/reference/pip_install/#requirements-file-format for more information

paho-mqtt==1.2